Parse a long format metabolic model to a glpk model

This function is deprecated. ROI.plugin.glpk is recommended instead.

expanded_to_glpk(reactions_expanded)

Arguments

reactions_expanded
A list of data frames as output by reactiontbl_to_expanded

Value

A list suitable for input to Rglpk

Details

This parses the long format produced by reactiontbl_to_expanded to a glpk model.

To install the Rglpk package in Linux, run sudo apt-get install libglpk-dev in a terminal, and then run install.packages('Rglpk') in R.

The reaction_table must have columns:

abbreviation,

equation,

uppbnd,

lowbnd, and

obj_coef.

See also