Parse a long format metabolic model to a Gurobi model

This function is deprecated. github.com/Fl0Sch/ROI.plugin.gurobi is recommended instead.

expanded_to_gurobi(reactions_expanded)

Arguments

reactions_expanded
A list of data frames as output by expand_reactions

Value

A list suitable for input to Gurobi.

Details

Used as the second half of reactiontbl_to_gurobi, this parses the long format produced by reactiontbl_to_expanded to a Gurobi model

For installation instructions for Gurobi, refer to the Gurobi website: http://www.gurobi.com/.

The reaction_table must have columns:

abbreviation,

equation,

uppbnd,

lowbnd, and

obj_coef.

See also