Merge metabolite stub and compartment to form an id

Merge metabolite stub and compartment to form an id

recompose_metabolites(expanded_metabolites)

Arguments

expanded_metabolites
a metabolite table as created by decompose_metabolites

Value

a merged metabolite table with one column, met

Examples

data(ecoli_core)

mod <- reactiontbl_to_expanded(ecoli_core)

decompose_metabolites(mod$mets)

#> # A tibble: 72 x 2
#>    chemical compartment
#>       <chr>       <chr>
#>  1    13dpg           c
#>  2      2pg           c
#>  3      3pg           c
#>  4     6pgc           c
#>  5     6pgl           c
#>  6    acald           c
#>  7    acald           e
#>  8       ac           c
#>  9    accoa           c
#> 10       ac           e
#> # ... with 62 more rows

recompose_metabolites(decompose_metabolites(mod$mets))

#> # A tibble: 72 x 1
#>         met
#>       <chr>
#>  1 13dpg[c]
#>  2   2pg[c]
#>  3   3pg[c]
#>  4  6pgc[c]
#>  5  6pgl[c]
#>  6 acald[c]
#>  7 acald[e]
#>  8    ac[c]
#>  9 accoa[c]
#> 10    ac[e]
#> # ... with 62 more rows