Apply gene expressions to reaction table

A convenience function that uses gene_eval and a custom function to apply new upper and lower bounds.

gene_associate(reaction_table, gene_table,
  expression_flux_function = function(x) {     (1 +
  log(x)/stats::sd(x)^2)^sign(x - 1) })

Arguments

reaction_table
A data frame describing the metabolic model.

gene_table
A data frame showing gene presence

expression_flux_function
a function to convert from gene set expression to flux

Value

the reaction_table, with a new column, present, and altered upper and lower bounds

Warning

This function relies on gene_eval, which uses eval to evaluate gene expression sets. This gives flexibility, but means that malicious code in the gene_sets argument could get evaluated. gene_sets is evaluated in a restricted environment, but there might be a way around this, so you might want to check for anything suspicious in this argument manually. For more information, read the code.

See also

gene_eval

Examples

data(iJO1366)
library(dplyr)

gene_table = data_frame(name = iJO1366$GRassoc %>%
stringr::str_split('and|or|\\s|\\(|\\)') %>%
  purrr::flatten_chr() %>%
  unique,
presence = 1) %>%
  filter(name != '', !is.na(name))

gene_associate(reaction_table = iJO1366 %>%
                 mutate(geneAssociation = GRassoc %>%
                          stringr::str_replace_all('and', '&') %>%
                          stringr::str_replace_all('or', '|')
                 ),
               gene_table = gene_table
)

#> # A tibble: 2,583 x 12
#>                      abbreviation
#>                             <chr>
#>  1                      DM_4CRSOL
#>  2                       DM_5DRIB
#>  3                      DM_AACALD
#>  4                        DM_AMOB
#>  5                      DM_MTHTHF
#>  6                        DM_OXAM
#>  7   Ec_biomass_iJO1366_WT_53p95M
#>  8 Ec_biomass_iJO1366_core_53p95M
#>  9                  EX_12ppd-R(e)
#> 10                  EX_12ppd-S(e)
#> # ... with 2,573 more rows, and 11 more variables: name <chr>, equation <chr>,
#> #   subsystem <chr>, lowbnd <dbl>, uppbnd <dbl>, obj_coef <dbl>,
#> #   GPRassoc <chr>, GRassoc <chr>, PRassoc <chr>, geneAssociation <chr>,
#> #   present <dbl>